ItFuture/PROGECTS/lantern2/backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-admin-tools==0.9.2
django-autoslug==1.9.9
django-cors-headers==3.14.0
django-filter==22.1
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.26.1
drf-yasg==1.21.5
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
MarkupSafe==2.1.2
nltk==3.8.1
oauthlib==3.2.2
packaging==23.0
Pillow==9.5.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.2
pywatchman==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
tqdm==4.65.0
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.15
webencodings==0.5.1
whitenoise==6.4.0
zipp==3.8.1
